middle-end: guard against non-single use compares in emit_cmp_and_jump_insns
When I wrote this optimization my patch stack included a change in
tree-out-of-ssa that would duplicate the compares such that the
use is always single use and get_gimple_for_ssa_name would always
succeed.

However I have dropped that for GCC 16 since I didn't expect the
vectorizer to be able to produce duplicate uses of the same
compare results.

But I neglected that you can get it by other means.  So this simply
checks that get_gimple_for_ssa_name succeeds for the LEN cases.
The non-LEN cases already check it earlier on.

To still get the optimization in this case the tree-out-of-ssa
change is needed, which is staged for next stage-1.

gcc/ChangeLog:

	* optabs.cc (emit_cmp_and_jump_insns): Check for non-single use.