Fix OMP SIMD clone mask record/get again
Post-CI checkin detected aarch64 fallout for the last change.  AARCH64
has ABI twists that run into a case where an unmasked call when loop
masked allows for a mask that has different shape than that of the
return value which in turn has different type than that of an actual
argument.

While we do not support a mismatch of call mask shape with the
OMP SIMD ABI mask shape when there's no call mask we have no such
restriction.

So the following fixes the record/get of a loop mask in the unmasked
call case, also fixing a latent issue present before.  In particular
do not record a random scalar operand as representing the mask.

A testcase is in gcc.target/aarch64/vect-simd-clone-4.c.

	* tree-vect-stmts.cc (vectorizable_simd_clone_call): Fix
	recording of the mask type again.  Adjust placing of
	mask arguments for non-masked calls.