fold-const, match.pd: Pass stmt to expr_not_equal if possible
The following patch is a small extension of the previous patch to pass stmt
context to the ranger queries from match.pd where possible, so that we can
use local ranges on a particular statement rather than global ones.

expr_not_equal_to also uses the ranger, so when possible this passes it
the statement context.

2025-11-27  Jakub Jelinek  <jakub@redhat.com>

	* fold-const.h (expr_not_equal_to): Add gimple * argument defaulted
	to NULL.
	* fold-const.cc (expr_not_equal_to): Likewise, pass it through to
	range_of_expr.
	* generic-match-head.cc (gimple_match_ctx): New static inline.
	* match.pd (X % -Y -> X % Y): Capture NEGATE and pass
	gimple_match_ctx (@2) as new 3rd argument to expr_not_equal_to.
	((A * C) +- (B * C) -> (A+-B) * C): Pass gimple_match_ctx (@3)
	as new 3rd argument to expr_not_equal_to.
	(a rrotate (bitsize-b) -> a lrotate b): Likewise.