ada: Spurious warning on untagged record type equality
The frontend reports a spurious warning when the type of some
component of an untagged record type is a private record type
that has an user-defined equality operator and the sources
are compiled enabling warnings for ignored equality operators
(that is, switch -gnatdw_q).

The warning is spurious because it must not be reported when
the full type declaration of the private type is a record type.

gcc/ada/ChangeLog:

	* exp_ch3.adb (Build_Untagged_Record_Equality): Add missing
	support for incomplete and private types.