gccrs: Partially unbreak rust build with C++20
I've committed earlier today https://gcc.gnu.org/r16-5628 to switch C++ to
-std=gnu++20 by default.  That apparently broke rust build (I don't have
cargo installed, so am not testing rust at all).

Here is a completely untested attempt to fix that.

Note, in C++20 u8"abc" literal has const char8_t[4] type rather than
const char[4] which was the case in C++17, and there is std::u8string
etc.
The casts below to (const char *) is what I've used in libcody as well
to make it compilable with all of C++11 to C++26.
Another thing is that the source for some reason expects -fexec-charset=
to be ASCII compatible and -fwide-exec-charset= to be UTF-16 or UTF-32
or something similar.  That is certainly not guaranteed.
Now, if rust-lex.cc can be only compiled with C++17 or later, we
could just use u8'_' etc., but as GCC still only requires C++14, I'd
go with u'_' etc.

2025-11-27  Jakub Jelinek  <jakub@redhat.com>

	* lex/rust-lex.cc (rust_input_source_test): Cast char8_t string
	literals to (const char *) to make it compilable with C++20.  Use
	char16_t or char32_t character literals instead of ordinary
	character literals or wide character literals in expected
	initializers.