optabs: Add else operand to LEN_LOAD.
When adding else operands to maskload and friends we didn't bother to do
the same for len_load (as we never use the residual elements anyway).
In order to simplify handling in gimple-fold, this patch adds the else
operand now.  Both, power and s390, zero out inactive elements.

gcc/ChangeLog:

	* config/rs6000/predicates.md (lxvl_else_operand): New
	predicate.
	* config/rs6000/vsx.md: Add else operand.
	* config/s390/predicates.md (vll_else_operand): New predicate.
	* config/s390/vector.md: Add else operand.
	* doc/md.texi: Document else operand.
	* internal-fn.cc (internal_fn_len_index): Adjust IFN_LEN_LOAD.
	(internal_fn_else_index): Add IFN_LEN_LOAD.
	* optabs-tree.cc (target_supports_len_load_store_p): Get else
	value for len_load.
	* tree-vect-stmts.cc (vectorizable_load): Pun the else value
	type.