doc: add mention to Algol 68 in the Installing GCC manual
The "Downloading the source" sectio in the "Installing GCC" manual
mentions all the compilers distributed within the GCC source
distribution.  This patch adds mentions to the ga68 compiler and its
libga68 run-time library.

Care is taken to describe the Algol 68 compiler as experimental,
according to the stipulations agreed with the SC.

Signed-off-by: Jose E. Marchesi <jemarch@gnu.org>

gcc/ChangeLog

	* doc/install.texi (Downloading the source): Mention Algol 68.