gfortran.texi: Remove spurious @menu entry
Fixed the issue:
  gfortran.texi:2542: warning: node up `Unsigned integers' in menu
  `Default exponents' and in sectioning `Extensions implemented in
  GNU Fortran' differ

There is an 'Unsigned integers' @menu entry under @section level
'Extensions implemented in GNU Fortran', where it should be. But some
spurious '@menu' entry, only with 'Unsigned integers' in it, was under
'@subsection Default exponents' just before the node of the
'@subsection Unsigned integers'. - The latter worked but was bogus
and lead the warning. Hence, it is now gone.

gcc/fortran/ChangeLog:

	* gfortran.texi (Default exponents): Remove spurious @menu entry.