[PATCH][PR target/122942] RISC-V: Add zifencei extension to the rva23s64 and rvb23s64 profiles [PR122942]
While verifying our Ascalon extensions list was complete (it isn't, follow-on
patch coming), I noticed that the rva23s64 and rvb23s64 profiles were missing
the mandatory zifencei extension, hence the following patch.

This was bootstrapped and regtested with no regressions.
Ok for trunk?

Peter

The RVA23S64 and RVB23S64 profiles both state that zifencei is a mandatory
extension.  Add it to both profiles.

gcc/
	PR target/122942
	* config/riscv/riscv-profiles.def (rva23s64): Add zifencei.
	(rvb23s64): Likewise.