c++/modules: Remove incorrect is_import check in using-directives
When I wrote this check in r16-5811 I was thinking of checking if the
using-directive was imported, but this just checks if the target
namespace was imported, which is not what we want.  We don't build deps
to see if the using-directive itself was imported, so just remove the
check.  I haven't been able to come up with a testcase this breaks but
it still seems reasonable to adjust.

gcc/cp/ChangeLog:

	* module.cc (module_state::write_using_directives): Correct
	condition.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>